refactor: api 명세 변경
