+++ backend/src/main/java/com/woowacourse/moamoa/study/controller/MyStudyController.java
@@ -6,6 +6,7 @@ import com.woowacourse.moamoa.study.service.response.MyStudiesResponse;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
@@ -20,5 +21,11 @@ public class MyStudyController {
     public ResponseEntity<MyStudiesResponse> getMyStudies(@AuthenticationPrincipal Long githubId) {
         final MyStudiesResponse myStudiesResponse = myStudyService.getStudies(githubId);
         return ResponseEntity.ok().body(myStudiesResponse);
+    }
+
+    @GetMapping("/api/members/me/role")
+    public ResponseEntity<Void> getStudyRole(
+            @AuthenticationPrincipal Long githubId, @RequestParam(name = "study-id") Long studyId) {
+        return ResponseEntity.ok().build();
     }
 }
